refactor(providers): consolidate app-wide providers under components/providers/
Move AvandarUiProvider and wrap PersistQueryClientProvider in a new
AvandarQueryClientProvider, both under a new src/components/providers/
directory. Migrations:

- src/components/AvandarUiProvider.tsx
  → src/components/providers/AvandarUiProvider.tsx
  (flat — no companion test or helper). All 12 import sites updated.

- src/lib/offline/queryPersister/queryPersister.{ts,test.ts}
  → src/components/providers/AvandarQueryClientProvider/queryPersister.{ts,test.ts}
  Colocated with the new provider that consumes it.

- New src/components/providers/AvandarQueryClientProvider/
  AvandarQueryClientProvider.tsx wraps PersistQueryClientProvider with
  the persister, 7-day maxAge, per-user buster, and
  shouldDehydrateQuery=success policy. Takes only `{ userId, children }`
  from the caller.

- src/main.tsx collapses the inline PersistQueryClientProvider block to
  `<AvandarQueryClientProvider userId={user?.id}>...`.

src/lib/offline/ is now down to `registerServiceWorker.ts`.
