+++ src/components/providers/AvandarQueryClientProvider/queryPersister.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from "vitest";
-import { makeCacheBuster } from "@/lib/offline/queryPersister/queryPersister";
+import { makeCacheBuster } from "@/components/providers/AvandarQueryClientProvider/queryPersister";
 
 describe("makeCacheBuster", () => {
   it("includes user id in the buster", () => {
