+++ src/test-utils/TestProviders.tsx
@@ -1,5 +1,5 @@
 import { I18nProvider } from "@lingui/react";
-import { AvandarUiProvider } from "@/components/AvandarUiProvider";
+import { AvandarUiProvider } from "@/components/providers/AvandarUiProvider";
 import { i18n } from "@/i18n/i18n";
 import type { ReactNode } from "react";
