+++ src/utils/testing-utils.ts
@@ -1,6 +1,6 @@
 import { RenderOptions, render as renderReact } from "@testing-library/react";
 import { ReactElement } from "react";
-import { AvandarUiProvider } from "@/components/AvandarUiProvider";
+import { AvandarUiProvider } from "@/components/providers/AvandarUiProvider";
 
 /**
  * Renders the given UI with the Avandar UI provider, which adds
